--- config/initializers/cors.rb
@@ -1,10 +1,0 @@
-# config/initializers/cors.rb
-Rails.application.config.middleware.insert_before 0, Rack::Cors do
-  allow do
-    origins 'http://127.0.0.1:3002' # Replace with your React app's URL
-    resource '*',
-      headers: :any,
-      methods: [:get, :post, :put, :patch, :delete, :options, :head],
-      credentials: true
-  end
-end
